refactor(cli): move android commands to their own module
